adding display update to the describing body after the target of an annotation was modified
If a user updated the selected text/target of an annotation, the annotation was correctly
updated in the wap-server and the elasticsearch-index. However, the displayed textCard and
the describing body (which holds the selected text as well) was not updated correctly.
Now the display of the annotation and describing body will be updated correctly.