adding new endpoint and changing the highlighting procedure of the textEditor
- creating a new endpoint "/{pageId}/displayableAnnotationsJSON", which serves
all displayable annotations of a page (annoJson). It behaves similar
to the "/{pageId}" endpoint, which adds the annoJson to the model.
- using the new endpoint to get an updated annoJson and then highlighting all
annotations after an annotation gets created/modified/deleted. This prevents
the existence of different states in the annoJson and the elasticSearchIndex/wap-server.
Previously the annoJson was updated on the client-side after each operation,
now the client just gets served an updated annoJson.
- minor changes to variables names in editor_text.html to improve visibility
of "global" (window.VARIABLENAME) variables.