changing the purpose of the body, which stores the type of annotation relevant for highlighting and linking
previously this information was stored in the "tagging" body, now it is stored in the "classifying" body. This
caused changes to multiple procedures and parts of the code
- on retrieval from the wap-server and inclusion into the elasticsearch index an annotation gets its color now
from the "classifying" body (AnnotationConverter.java)
- on annotation creation in the browser, now "classifying" will be used as the purpose for the body, where
"mrw direct", "metaphor" etc. is getting stored (creation_templates_text.js)
- for metaphor-annotation creation the way of storing mrw-annotations included in the current selection changed
(editor_xml.js and creation_templates_text.js)
- adding "classifying"-bodies to be properly displayd in the horizontal textCard (editor.js)