Remove obsolete @svgdotjs from node_modules folder
Even when svgdotjs was used for the first time (and installed into
node_modules/), the local version was not used at all. The library
is instead loaded from a CDN. But since it was checked into version
control, it is still present on every checkout, even though node_modules/
is supposed to be ignored by git.