include modularized navbar for texteditor
This commit includes modularized js source code for the texteditor,
which serves as a drop-in replacement for generation of the chapter
nav bar. It enables the js/texteditor.js entry point in the webpack
build pipeline. Last but not least, it adjusts the editor_text.html
template to use the new solution instead of the previous one.