make concept search logging less intrusive
We now no longer store the datetime of the query, but rather an
arbitrary string to indicate that multiple queries have been issued
in a single sitting. This string is re-calculated on every page load
(that is, all the time, even during one "session").
Also, we now only log search queries when/if the annotation is actually
saved. This way, it is possible to play around with the concept linking
without something getting logged.
Fixes https://git.noc.ruhr-uni-bochum.de/sfb1475-inf/takita/-/issues/99