changing retrieval of pageId, making a function use parameters and updating documentation
the retrieval of the pageId from the current location was implemented a long time ago.
It was unreliable using a contextpath so it was changed. As the pageId is available
for the textEditor as a global variable, there is no need to use the location anymore.
Now the function can be tested (jest doesn't implement the full navigation functionality,
so the retrieval of the pageId from the location was impossible in the test-environment).