fixing displaying an annotation selected via the annotation tables (overview and annotator)
- both navigation bars will be initialized and will show the necessary division first;
only afterwards scroll to the target of the annotation. The initialization of the
second level navigation bar does no longer care about scrolling to the target of the
annotation as the initialization of the first level navigation bar takes care of it.
Both initializations take care of displaying the correct divison for ther respective levels.
- the function to get the div-element, which contains the target of an annotation,
is now more robust as it stops when encountering the text container element
- updating documentation