changing annotation target highlighting to improve the display of the selected annotation
by default all annotation targets will be highlighted like before (in a normal
version of a color). If a user selects one annotation, only the targets of this
annotation will keep the normal version of the color; the targets of all other
annotations will be highlighted in a lighter color. Unselecting the annotation
will highlight all annotations in a normal version of the color again.