fixing en/disabling of authentication via application property
the property "takita.security.enabled" is used to determine whether authentication is
enabled and which filterChain should be used. The CSRF token is passed to the frontend
via thymeleaf and stored in a meta-element; if "takita.security.enabled" is set to false
CSRF protection is disabled and the frontend uses a dummy value for the HTTP requests.